feat(engine): server endpoint validation on initialization
RBroker instances accept but do not validate or test the DeployR server endpoint
passed on their RBrokerConfig. For anonymous RBroker instances this means
invalid server endpoints are not detected until the first attempt to execute an
RTask, potentially long after RBroker initialization has completed. This is not
conducive to building robust client applications.

Authenticated RBroker instances use the server endpoint at initialization
(in an attempt to authenticate with the server) so are not prone to this issue.